chore: extract posts to a seperate file
